ci: add Docker-based build and release job
Add A Docker-based build and release job. This uses an upstream workflow
along with a multi-stage Dockerfile to build an image with the ref's
binary.

If triggered via PR or push to main, this pushes a pre-release image to
the registry.

If triggered via tag, this pushes a release image and creates a Github
release.

Signed-off-by: Travis Raines <571832+rainest@users.noreply.github.com>